requirements.txt
transformers
datasets
sentence-transformers
langchain
chromadb
huggingface-hub
ragas
streamlit
pandas
python-dotenv
ipykernel
tqdm
torch
langchain-groq
langchain-huggingface
ragas
datasets
evaluate
langchain-chroma 
